+++ CHANGELOG.md
@@ -1,2 +1,2 @@
 ## Change
-- sync with `sakura-ryoko/itemscroller` 1.21-0.24.57
+- update MaFgLib, update NeoForge version to `21.1.146`